Refactor Show.py -> show.py
Refactor Show.py -> show.py
