+++ medusa/helpers.py
@@ -60,7 +60,7 @@ from . import classes, db
 from .common import USER_AGENT
 from .helper.common import episode_num, http_code_description, media_extensions, pretty_file_size, subtitle_extensions
 from .helper.exceptions import ex
-from .show.Show import Show
+from .show.show import Show
 
 
 logger = logging.getLogger(__name__)
