+++ pytest.ini
@@ -199,7 +199,7 @@ flake8-ignore =
     medusa/server/web/manage/searches.py D100 D101 D102 N802
     medusa/show/ComingEpisodes.py D100 D102 D202 D204 D205 D400
     medusa/show/History.py D100 D101 D102 D200 D202 D204 D205 D400 D401
-    medusa/show/Show.py D100 D101 D102 D202 D205 D400
+    medusa/show/show.py D100 D101 D102 D202 D205 D400
     medusa/show/__init__.py D104
     medusa/show/recommendations/__init__.py D104
     medusa/show/recommendations/anidb.py D100 D101 D400
